+++ src/bunit.web/Rendering/Internal/Htmlizer.cs
@@ -15,8 +15,7 @@ namespace Bunit
 	/// This file is based on
 	/// https://source.dot.net/#Microsoft.AspNetCore.Mvc.ViewFeatures/RazorComponents/HtmlRenderer.cs
 	/// </summary>
-	[SuppressMessage("Usage", "BL0006:Do not use RenderTree types", Justification = "<Pending>")]
-	internal class Htmlizer
+	internal static class Htmlizer
 	{
 		private static readonly HtmlEncoder HtmlEncoder = HtmlEncoder.Default;
 
@@ -27,22 +26,25 @@ namespace Bunit
 
 		private const string BLAZOR_INTERNAL_ATTR_PREFIX = "__internal_";
 		private const string BLAZOR_CSS_SCOPE_ATTR_PREFIX = "b-";
-		public const string BLAZOR_ATTR_PREFIX = "blazor:";
-		public const string ELEMENT_REFERENCE_ATTR_NAME = BLAZOR_ATTR_PREFIX + "elementreference";
+		internal const string BLAZOR_ATTR_PREFIX = "blazor:";
+		internal const string ELEMENT_REFERENCE_ATTR_NAME = BLAZOR_ATTR_PREFIX + "elementreference";
 
 		public static bool IsBlazorAttribute(string attributeName)
-			=> attributeName.StartsWith(BLAZOR_ATTR_PREFIX, StringComparison.Ordinal) ||
-			   attributeName.StartsWith(BLAZOR_CSS_SCOPE_ATTR_PREFIX, StringComparison.Ordinal);
+		{
+			if (attributeName is null) throw new ArgumentNullException(nameof(attributeName));
+			return attributeName.StartsWith(BLAZOR_ATTR_PREFIX, StringComparison.Ordinal) ||
+						  attributeName.StartsWith(BLAZOR_CSS_SCOPE_ATTR_PREFIX, StringComparison.Ordinal);
+		}
 
 		public static string ToBlazorAttribute(string attributeName)
 		{
 			return $"{BLAZOR_ATTR_PREFIX}{attributeName}";
 		}
 
-		public static string GetHtml(ITestRenderer renderer, int componentId)
+		public static string GetHtml(int componentId, RenderTreeFrameCollection framesCollection)
 		{
-			var frames = renderer.GetCurrentRenderTreeFrames(componentId);
-			var context = new HtmlRenderingContext(renderer);
+			var context = new HtmlRenderingContext(framesCollection);
+			var frames = context.GetRenderTreeFrames(componentId);
 			var newPosition = RenderFrames(context, frames, 0, frames.Count);
 			Debug.Assert(newPosition == frames.Count, $"frames.Count = {frames.Count}. newPosition = {newPosition}");
 			return string.Join(string.Empty, context.Result);
@@ -101,7 +103,7 @@ namespace Bunit
 			int position)
 		{
 			ref var frame = ref frames.Array[position];
-			var childFrames = context.Renderer.GetCurrentRenderTreeFrames(frame.ComponentId);
+			var childFrames = context.GetRenderTreeFrames(frame.ComponentId);
 			RenderFrames(context, childFrames, 0, childFrames.Count);
 			return position + frame.ComponentSubtreeLength;
 		}
@@ -264,12 +266,15 @@ namespace Bunit
 
 		private class HtmlRenderingContext
 		{
-			public ITestRenderer Renderer { get; }
+			private readonly RenderTreeFrameCollection _frames;
 
-			public HtmlRenderingContext(ITestRenderer renderer)
+			public HtmlRenderingContext(RenderTreeFrameCollection frames)
 			{
-				Renderer = renderer;
+				_frames = frames;
 			}
+
+			public ArrayRange<RenderTreeFrame> GetRenderTreeFrames(int componentId)
+				=> _frames[componentId];
 
 			public List<string> Result { get; } = new List<string>();